Changed double to float for price

+++ src/models/binance_model.py
@@ -1,6 +1,5 @@
 from datetime import datetime
 
-from numpy import double
 from pydantic import BaseModel
 
 
@@ -18,6 +17,6 @@ class BinanceRawData(BaseModel):
 
 class BinanceTransformedData(BaseModel):
     symbol: str
-    price: double
+    price: float
     time: datetime
     source: str
